TASK: Flow 5 compatibility (#4)
* TASK: Apply migration Neos.SwiftMailer-20161130105617

Adjusts code to package renaming from "TYPO3.SwiftMailer" to
"Neos.SwiftMailer".

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Fusion-20180211175500

Expand Neos.Neos-FusionObjects without namespace to fully qualified
names

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Fusion-20180211184832

Remove namespace aliases and replace them with the package keys

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20180415105700

Add scalar type hint to CacheAwareInterface implementations.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Neos-20180907103800

Migrate RTE formatting configuration to new format

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Added Flow 5 compability